+++ src/main/java/com/groom/orbit/goal/app/dto/request/UpdateQuestSequenceRequestDto.java
@@ -1,3 +1,3 @@
-package com.groom.orbit.goal.app.dto;
+package com.groom.orbit.goal.app.dto.request;
 
 public record UpdateQuestSequenceRequestDto(Long questId, int sequence) {}
