+++ src/main/java/com/groom/orbit/goal/app/dto/response/GetGoalCategoryResponseDto.java
@@ -1,4 +1,4 @@
-package com.groom.orbit.goal.app.dto;
+package com.groom.orbit.goal.app.dto.response;
 
 import java.util.List;
