+++ src/main/java/com/groom/orbit/goal/app/dto/response/GetOnGoingGoalResponseDto.java
@@ -1,4 +1,4 @@
-package com.groom.orbit.goal.app.dto;
+package com.groom.orbit.goal.app.dto.response;
 
 public record GetOnGoingGoalResponseDto(
     Long goalId, String title, long totalQuestCount, long finishQuestCount) {}
