+++ src/main/java/com/groom/orbit/goal/app/dto/response/QuestInfoResponseDto.java
@@ -1,3 +1,3 @@
-package com.groom.orbit.goal.app.dto;
+package com.groom.orbit.goal.app.dto.response;
 
 public record QuestInfoResponseDto(Long id, String title, Boolean isComplete) {}
