+++ src/main/java/com/groom/orbit/goal/app/query/GoalQueryService.java
@@ -3,7 +3,7 @@ package com.groom.orbit.goal.app.query;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.groom.orbit.goal.app.dto.GetGoalCategoryResponseDto;
+import com.groom.orbit.goal.app.dto.response.GetGoalCategoryResponseDto;
 import com.groom.orbit.goal.dao.entity.GoalCategory;
 
 import lombok.RequiredArgsConstructor;
