+++ src/main/java/com/groom/orbit/goal/controller/query/GoalQueryController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.goal.app.dto.GetGoalCategoryResponseDto;
+import com.groom.orbit.goal.app.dto.response.GetGoalCategoryResponseDto;
 import com.groom.orbit.goal.app.query.GoalQueryService;
 
 import lombok.RequiredArgsConstructor;
