+++ src/main/java/com/groom/orbit/goal/controller/query/QuestQueryController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.ResponseDto;
-import com.groom.orbit.goal.app.dto.QuestInfoResponseDto;
+import com.groom.orbit.goal.app.dto.response.QuestInfoResponseDto;
 import com.groom.orbit.goal.app.query.QuestQueryService;
 
 import lombok.RequiredArgsConstructor;
